requirements.txt
# Core GUI
PySide6>=6.5.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# AI/ML
deepface>=0.0.79
rembg>=2.0.38

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0
pyinstaller>=5.13.0

# Windows specific (for printing)
pywin32>=305

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0